+++ Reynj/Linq/Inverse.cs
@@ -55,6 +55,9 @@ namespace Reynj.Linq
         {
             if (source == null) 
                 throw new ArgumentNullException(nameof(source));
+            if (minValue.CompareTo(maxValue) > 0)
+                throw new ArgumentException($"{nameof(maxValue)} must be greater than or equal to {nameof(minValue)}",
+                    nameof(maxValue));
 
             var reducedSource = source
                 .Reduce() // Sorts the collection, merges overlapping and touching ranges and removes empty ranges